Moved more hook code into uhook.cpp and created an inner class just to get the char * for a FName value
