+++ CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 2.6)
+project(CMAKEDEMO)
+
+add_subdirectory(unrealspy)
+
+#cmake -G "NMake Makefiles" .
